FEATURE: `prox-dush` now checks for LXC containers, and prints usage
changes to `prox-dush` in this commit include:
- the script now checks for containers in `/etc/pve/lxc`
- in addition, the script does "usage calculations"

for the "usage calculations", the script prints a percentage value for each object (the "usage %"). the "usage %" is equal to the size of the VM or container, divided by the total used space of VMs and containers. in other words, it is not the "usage %" compared to the entire disk.